refactor: Separate state from presentation
